.wip(structure): splitted code
